Check status of install
